Move input properties into state machine.
